+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* Add a `password-alignment` option to customize the alignment of the text in the
+  password input.
 * Add an `invalid-password-text` option to customize the wrong password error
   text.